Rename forked dq_* file to kqp_*
