+++ ydb/core/kqp/opt/cbo/solver/kqp_opt_join.cpp
@@ -1,4 +1,4 @@
-#include "dq_opt_join.h"
+#include "kqp_opt_join.h"
 #include <ydb/library/yql/dq/opt/dq_opt_phy.h>
 #include <yql/essentials/core/yql_cost_function.h>
