+++ ydb/core/kqp/opt/cbo/solver/kqp_opt_make_join_hypergraph.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "dq_opt_join_hypergraph.h"
-#include "dq_opt_conflict_rules_collector.h"
+#include "kqp_opt_join_hypergraph.h"
+#include "kqp_opt_conflict_rules_collector.h"
 
 #include <library/cpp/disjoint_sets/disjoint_sets.h>
 #include <ydb/core/kqp/opt/cbo/cbo_optimizer_new.h>
