+++ ydb/core/kqp/opt/cbo/solver/kqp_opt_predicate_selectivity.cpp
@@ -1,4 +1,4 @@
-#include "dq_opt_stat.h"
+#include "kqp_opt_stat.h"
 
 #include <yql/essentials/core/yql_opt_utils.h>
 #include <yql/essentials/utils/log/log.h>
