+++ ydb/core/kqp/opt/kqp_statistics_transformer.h
@@ -11,7 +11,7 @@
 #include <yql/essentials/core/yql_expr_type_annotation.h>
 #include <ydb/core/kqp/provider/yql_kikimr_provider_impl.h>
 #include <yql/essentials/core/yql_opt_utils.h>
-#include <ydb/core/kqp/opt/cbo/solver/dq_opt_stat_transformer_base.h>
+#include <ydb/core/kqp/opt/cbo/solver/kqp_opt_stat_transformer_base.h>
 
 namespace NKikimr {
 namespace NKqp {
